requirements.txt
# Eliminado el certificado (a IV no le funciona)
# Rebajada la version de gym a 0.21.0
# Rebajada la version de PettingZoo a 1.19.0 (Junio 2022) y compatible con gym==0.21.0


absl-py==1.2.0
ale-py==0.7.4
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.1.0
cycler==0.11.0
fonttools==4.36.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
gym==0.21.0
gym-notices==0.0.8
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
PettingZoo==1.19.0
Pillow==9.2.0
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stable-baselines3==1.6.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
zipp==3.8.1
networkx